requirements.txt
amqp               ==5.1.1
async-timeout      ==4.0.2
billiard           ==3.6.4.0
blinker            ==1.5
celery             ==5.2.7
certifi            ==2022.9.24
charset-normalizer ==2.1.1
click              ==8.1.3
click-didyoumean   ==0.3.0
click-plugins      ==1.1.1
click-repl         ==0.2.0
clipboard          ==0.0.4
commonmark         ==0.9.1
Deprecated         ==1.2.13
Flask              ==2.2.2
Flask-PyMongo      ==2.3.0
idna               ==3.4
itsdangerous       ==2.1.2
Jinja2             ==3.1.2
kombu              ==5.2.4
MarkupSafe         ==2.1.1
packaging          ==21.3
pip                ==22.2.2
prompt-toolkit     ==3.0.31
Pygments           ==2.13.0
pymongo            ==4.2.0
<<<<<<< HEAD
=======
pymysql            ==1.0.2
>>>>>>> add utils module
pyparsing          ==3.0.9
pyperclip          ==1.8.2
python-dotenv      ==0.21.0
pytz               ==2022.4
redis              ==4.3.4
requests           ==2.28.1
rich               ==12.5.1
setuptools         ==62.3.2
six                ==1.16.0
urllib3            ==1.26.12
vine               ==5.0.0
wcwidth            ==0.2.5
Werkzeug           ==2.2.2
wheel              ==0.37.1
wrapt              ==1.14.1
